Admin app. Module style refactoring
