--- angular-spa-ui/js/admin-app/content-manager-module/cm-dir.js
@@ -1,5 +1,0 @@
-"use strict";
-
-module.exports = function (cm) {
-
-}
