+++ angular-spa-ui/js/admin-app/content-manager-module/contentManager-config.js
