+++ angular-spa-ui/js/admin-app/content-manager-module/contentManager-dir.js
@@ -2,4 +2,4 @@
 
 module.exports = function () {
 
-};
+}
