+++ angular-spa-ui/js/admin-app/content-manager-module/contentStorage-srv.js
@@ -1,0 +1,8 @@
+module.exports = function () {
+    return {
+        data: {},
+        params: {},
+        details: {},
+        searchType: {}
+    }
+};
