+++ angular-spa-ui/js/admin-app/content-manager-module/recordsListDir/recordsList-dir.js
@@ -1,43 +1,39 @@
 "use strict";
 
-module.exports = function (app) {
+module.exports = function (rlService, configService) {
+    return {
+        restrict: "E",
+        scope: {
+            items: '=itemslist',
+            header: '=headerSettings',
+            itemsConfig: '=itemsSettings',
+        },
+        templateUrl: '/results.html',
+        controller: function ($scope) {
 
-    app.directive('recordsList', function (rlService, configService) {
-        return {
-            restrict: "E",
-            scope: {
-                items: '=itemslist',
-                header: '=headerSettings',
-                itemsConfig: '=itemsSettings',
-            },
-            templateUrl: '/results.html',
-            controller: function ($scope) {
+            $scope.model = rlService.setDefaultModel();
 
-                $scope.model = rlService.setDefaultModel();
+            // Records List actions
+            $scope.goToPage = function () {
+                $scope.$emit('goToPage', $scope.model.currentPage);
+            };
 
-                // Records List actions
-                $scope.goToPage = function () {
-                    $scope.$emit('goToPage', $scope.model.currentPage);
-                };
+            $scope.sortChange = function () {
+                $scope.$emit('SetSortBy', $scope.model.sortBy.value);
+            };
 
-                $scope.sortChange = function () {
-                    $scope.$emit('SetSortBy', $scope.model.sortBy.value);
-                };
+            $scope.limitChange = function () {
+                $scope.$emit('SetLimit', $scope.model.limit.value);
+            };
 
-                $scope.limitChange = function () {
-                    $scope.$emit('SetLimit', $scope.model.limit.value);
-                };
+        },
+        link: function (scope, element, attrs) {
 
-            },
-            link: function (scope, element, attrs) {
+            scope.$watchCollection('header', function (newValue, oldValue) {
+                if (!_.isEqual(oldValue, newValue))
+                    scope.model = rlService.setModelValues(newValue, scope.model);
+            })
 
-                scope.$watchCollection('header', function (newValue, oldValue) {
-                    if (!_.isEqual(oldValue, newValue))
-                        scope.model = rlService.setModelValues(newValue, scope.model);
-                })
-
-            }
         }
-    });
-
-}
+    }
+};
